requirements.txt
# Discord library (Python 2 compatible)
discord.py==0.16.12

# Environment variables
python-dotenv==0.19.2

# HTTP requests for webhook support
requests==2.31.0

# Optional: for compatibility with old Python 2 features
future==0.18.3
colorama==0.4.6
